I now can create a file if the file doesn't exist

+++ Day-30-Errors-Exceptions-JSON/Day-29-PM-JSON/main.py
@@ -2,6 +2,7 @@ from tkinter import *
 from tkinter import messagebox as mb
 from password_gen import random_password_generator
 import json
+from pathlib import Path
 
 RANDOM_PASSWORD = None
 window = Tk()
@@ -25,18 +26,36 @@ def add_to_file():
         mb.showinfo("Opps", "Please make sure you didn't leave anything empty")
     else:
         if mb.askyesno('Final Step', 'You sure the data are correct?'):
-            with open('password.json', mode='r') as password_file:
-                json_list = json.load(password_file)
-            with open('password.json', mode='w') as password_file:
-                json_list.append(obj_to_save)
-                json.dump(json_list, password_file,indent=4)
-            email_entry.delete(0, 'end')
-            password_entry.delete(0, 'end')
-            website_entry.delete(0, 'end')
+
+            if Path('password.json').is_file():
+                with open('password.json', mode='r') as password_file:
+                    data = json.load(password_file)
+                    data.update(obj_to_save)
+                with open('password.json', mode='w') as password_file:
+                    json.dump(data, password_file, indent=4)
+            else:
+                with open('password.json', mode='w') as password_file:
+                    json.dump(obj_to_save, password_file, indent=4)
+
+
+
+            # with open('password.json', mode='w') as password_file:
+            #     json_list.append(obj_to_save)
+            #     json.dump(json_list, password_file,indent=4)
+            # email_entry.delete(0, 'end')
+            # password_entry.delete(0, 'end')
+            # website_entry.delete(0, 'end')
 
 
 def search_in_file():
-    pass
+    website = website_entry.get().lower()
+    with open('password.json', mode='r') as password_file:
+        json_list = json.load(password_file)
+    try:
+        website_keys = [key.keys() for key in json_list]
+        print('facebook' in website_keys)
+    except:
+        pass
 
 
 logo = Canvas(height=224, width=200, highlightthickness=0)
